+++ lib/enumerize/hooks/formtastic.rb
@@ -11,8 +11,13 @@ module Enumerize
 
       def input_with_enumerize(method, options={})
         klass = object.class
+
         if klass.respond_to?(:enumerized_attributes) && (attr = klass.enumerized_attributes[method])
           options[:collection] ||= attr.options
+
+          if attr.kind_of?(Enumerize::Multiple) && options[:as] != :check_boxes
+            options[:input_html] = options.fetch(:input_html, {}).merge(:multiple => true)
+          end
         end
 
         input_without_enumerize(method, options)